Improved parfor optimizer (awareness of cp mem of spark instructions)
This changes adds awareness wrt additional cp memory requirements of
spark instructions to the parfor optimizer. In detail this includes two
changes: (1) awareness of broadcast cp memory requirements, and (2) a
more conservative decision on the degree of parfor parallelism in order
to prevent unnecessary guarded collect. These changes are important for
for robustness in terms of runtime and memory requirements.