Performance multi-threaded wdivmm left (col parallelization), for als_cg
So far both wdivmm left and right were parallelized over rows in W for
multi-threaded environments. This strategy performed very well for
wdivmm right due to disjoint outputs. However, for wdivmm left it
required thread-local outputs and final aggregation. Especially for
large factors, this was very cache-unfriendly because it scanned the
entire output for each row in W. In order to overcome this performance
bottleneck, this change reimplements wdivmm left with a column-wise
parallelization strategy. Even for factors of 800MB and very sparse
datasets (up to 10^{-7}), this already showed a 3x improvement. Finally,
wdivmm right still uses a row-wise parallelization strategy.