Performance wdivmm minus patterns (nz-indicator awareness), for als_cg
In the initial version of wdivmm -minus patterns we supported the
general case of arbitrary weights which are not necessarily non-zero
indicators. Especially, for the spark/mr backends this generality led to
a severe performance bottleneck due to repeated shuffling of X and W.
This change restricts the wdivmm minus patterns to weights of non-zero
indicators which allows data-local map-side computation.