Improved parfor optimizer rewrite 'degree of parallelism', for kmeans
The parfor degree of parallelism for local parfor is bounded by the
problem size and the fraction of memory budget and memory estimates (in
the presence of spark instructions even more conservative). Any
remaining parallelism is then recursively distributed to nested parfors
and operations. This approach was too conservative since our
multi-threaded operations do not have additional memory requirements.
With this change, we now separately assign parfor and operation degree
of parallelism. For example, consider a case w/ 24 vcores and large
memory requirements that led to parfor k=1 - so far, this resulted in
k=1 for any body operation as well, while with this change we now assign
k=24 to body operations. This is especially important for
compute-intensive cases like kmeans.