Improved right indexing hop size inference (block indexing), for kmeans
The parfor optimizer injects repartition operations if there are zipmm
or cpmm (unknown) matrix mult instructions in the body in order to avoid
shuffle per iteration. For kmeans, we encountered false positives, where
we injected repartition operations but executed mapmm in the body, which
rendered the repartition overhead unnecessary. This change improves the
size inference logic of right indexing hops such that, for kmeans, all
sizes of matrix multiplications can be inferred and thus avoids
unnecessary repartitioning.