New cp unary/binary guarded dense/sparse output conversion, for kmeans
By default we estimate the output sparsity on each operation, allocate
the right output, execute the operation, and only check for explicit
dense/sparse conversion if the optimizer reflects this in the memory
estimates used for operator selection. This leads to issues with certain
binary/unary operations, resulting in wrong output format (e.g.,
mistakenly dense) and thus affects performance of subsequent operations.
For example, P=ppred(D, minD, "<="); is estimated as dense although it
should be sparse for sufficiently many clusters. This change adds a
guarded sparse/dense format check (incl conversion if necessary) to all
binary and unary cp instructions. The additional memory requirements of
a potential format change are guarded by the released memory of all
inputs (if sufficiently large).