New rdd block partitioner, incl aggregation utils; still disabled
Local aggregation is crucial for performance by reducing the amount of
shuffled data. Especially for squared sparse matrices (as common for
factorization and graph algorithms), the default hash partitioning leads
to "randomly" distributed blocks and hence, almost no local aggregation
for row or column-wise aggregation. This custom partitioner assigns
squared matrix tiles that are composed of multiple blocks to individual
partitions and hence ensures local aggregation independent of the
aggregation direction. 

With this first change, we introduce the partitioner and extended
aggregation utils (which require the matrix dimensions). Down the road,
this will become the default binary block partitioner if the promising
initial results hold over a variety of additional experiments. Note that
all shuffle-based operations would need to change simultaneously in
order to avoid unnecessary overhead due to partitioner mismatches.