[SYSTEMML-157][SYSTEMML-300] Fix python parser (cast function handling)
This change fixes a general issue of our python parser that led to not
recognizing cast functions (after mapping to dml names) with '.' in
their names as builtin functions. Due to this problem, any as.scalar,
as.matrix, as.integer, as.logical, as.double in assignment statements
were incorrectly dispatched to functional calls. The fix is a simple
whitelist for now.