[SYSTEMML-382] Fix consistency sparse row deep-copy on row inserts
Especially with the new update in-place support we have to be careful
about shallow copies of deep rows, even if the current operation does
not modify the physical representation. This patch makes three
improvements. First, it fixes several issues introduces with the sparse
block runtime integration. Second, it extends the set row API by a flag
for deep copy requirements, which prevents unnecessary external deep
copies in case of CSR and COO. Third, it improves the performance of row
appends and specific binary-scalar operations by avoiding repeated
sparse row re-allocations. 

With this fix we can ensure result correctness again but might be too
conservative. During the rework of update in-place, we need to revisit
this and define a general policy of sparse row allocations.