Replace SyntacticErrorListener with CustomErrorListener, step 2
Stage 2 of the replacement of SyntacticErrorListener with its inner
class CustomErrorListener.  In order to retain history, we do this
in two commits: one commit to change the name of the
SyntacticErrorListener class to CustomErrorListener while changing
the inner CustomErrorListener class to a dummy name; a second commit
to move the inner class methods out to the outer class.

Closes #61, Closes #73.