[SYSTEMML-510] Fix performance sparse wdivmm with scalar epsilon
The newly added wdivmm w/ eps always selected the generic fallback
implementation which led to severe performance issues for the common
case of sparse-dense-dense. In addition to the fixed selection of
primitives, this patch also includes a fix for the sparse-dense-dense
implementation with scalars (which ran into null pointer exceptions).