[SYSTEMML-555] Transform builtin cleanup (spec as string argument)
So far transform took the 'transform specification' as a json file,
where the file path was an input argument. This approach is very
inflexibile with regard to CP operations and jmlc scoring, where we aim
for pure in-memory computation. This patch changes the transform
definition to take the json string as an input argument - now the
specification can be either read from a scalar file or constructed in
memory and passed to transform. Note, however, that we are not yet able
to construct the json specification in DML via string concatenation due
to missing functionality for string escaping wrt quotes.