[SYSTEMML-556] JMLC api extension for input and output frames
This patch adds input and output frames to the JMLC API. Frames are
specified by a mixed schema or implicit string schema and data is passed
as a two dimensional string array or directly as a FrameBlock.
Furthermore, this change also includes extended casting functionalities
between frames and string arrays.