[SYSTEMML-556] JMLC support for frame column names, incl jmlc test
As a preparation step for modifying transform to work over frames, we
need to expose frame column names through the jmlc api and carry this
information as part of frames. Furthermore, this patch also includes an
initial jmlc transform test (but is still disabled until the full rework
of transform happened).