[SYSTEMML-555] New transformapply builtin function over cp frames
This patch introduces a new transformapply builtin function over
in-memory frames consuming both input and metadata as frames. In order
to allow for a smooth transition from the existing transform, it exists
redundantly to the transform apply functionality. Once we have full
bufferpool integration and distributed operations for transformapply, we
will remove the old apply functionality from transform. Note that right
now transformapply is limited to recoding. Finally, this change also
includes a fix for frames (maintaining the number of rows) as well as a
transformapply jmlc test.