[SYSTEMML-567] Initial buffer pool integration of frames, incl cleanup
This patch realizes an initial integration of frames into our buffer
pool (caching framework). However, note that frames are not yet subject
to eviction. In detail, this includes (1) a generalization of the base
class for cacheable objects (frame and matrix), (2) execution context
primitives for obtaining/returning frames, and (3) exploitation in
operations over frames (transformapply, transformdecode) as well as APIs
(jmlc). 

Furthermore, this change also includes various cleanups of exception
handling and state management as well as initial hooks for frame readers
and writers.