[SYSTEMML-584] New thread-local configuration handling (dml/compiler)
So far we did not fully support concurrent script executions in a shared
JVM process (e.g., as necessary for JMLC scoring in spark executors)
because global static state updates caused potential side effects. This
patch introduces thread-local configuration handling for dml and
compiler configurations. All these functionalities are encapsulated
within the ConfigurationManager behind a very generic API. This cleanup
covers all relevant configurations for JMLC - down the road we could
extend this as required. Note that we also support 1-1 thread mappings
in parfor, which will be required in JMLC for parfor scripts with
disabled parallelism.