[SYSTEMML-541] Fix reblock recompile test (seed and input data)
The occasional issues with this test were due to different definitions
of our ctable compared to R's table that show up whenever certain groups
are not represented at all. So far the input data was generated as 20x1
from a [1,5] domain (5 distinct values for ctable and grouped
aggregate). This patch makes two improvements: (1) we now test against
both random and fixed seed, and (2) use an input matrix of 2000x1 which
makes differences *very* unlikely.