[SYSTEMML-396] New multi-threaded cumulative aggregates (e.g., cumsum)
This patch introduces multi-threaded cumulative aggregates, which is
especially useful for cumsum because the used numerically stable kahan
plus is rather costly. However, the speedup is limited as a significant
fraction of time is spent in output allocation. In experiments over both
dense and sparse, we have seen speedups of 1.5x-3x, depending the used
JDK.