[SYSTEMML-533] Fix variable name placeholder for instruction patching
Whenever instructions expect constant value arguments but we have
variables during compilation, we generate placeholders and rely on
instruction patching during runtime. So far we used ## which runs into
collisions if used as literals. This patch changes this placeholder to
the extended ASCII character \u00b6 to avoid collisions and cleans up
remaining code that directly referred to the placeholder instead of the
global compiler constant.