Cache-conscious sparse-sparse core block matrix multiplication
In a similar spirit as the new sparse-dense cache-conscious matrix
multiply, this patch also makes sparse-sparse matrix multiply
cache-conscious. However, due to the sparse inner loop, we use a best
effort cache blocking with dynamic block sizes computed from the
sparsity of A to reflect average row reuse. Overall, we've seen
performance improvements of up to 2x without degradation over a variety
of usecases with different sparsity and shapes.