[SYSTEMML-399] Performance parallel textcell reader (nnz, sort sparse)
Similar to parallel binary block readers, this patch improves the
parallel textcell reader to do any post-processing (nnz computation and
sorting of sparse rows) in parallel in order to reduce the serial
fraction of the entire read. This improvement is always beneficial but
especially important on many core machines with a large number of
threads. Furthermore, this change also cleans up the brittle error
handling via proper handling of task futures as well as fixes an issue
with the consolidated buffer pool integration of frames.