Improved mr remove empty compilation (broadcast-based pmm if possible)
This patch introduces advanced operator selection for MR remove empty
row operations. Similar to SYSTEMML-215, we employ a broadcast-based
operation if the selection vector fits into the broadcast. However,
instead of introducing a new physical operator, we directly compile a
PMM operation for simplification and consistency (across alternative
formulations of row selection) because this operator has precisely the
semantics of the actual selection step in remove empty. On a scenario of
250GB, this change improved performance from 1,587s (4 jobs) to 573s (3
jobs) by preventing an expensive shuffle-heavy job altogether.