[SYSTEMML-495] Simplify SystemML Configuration Loading
Always looking for a default file, even when a file is explicitly supplied, causes confusion, and may create unintended side effects.  If a user explicitly supplies a configuration file, we should simply use that file, rather than attempting to also read an additional default file at the same time.  In order to provide backwards compatibility, if the user does not supply a file, then we should fall back to the default file location.  If that file is also not found, we should then fall back to the internal defaults.

Additionally, in order to supply an additional configuration file, the `-config=path/to/file` flag has to be used, and it only works with the equals sign `=` present, which is unlike all other flags that use spaces.  Instead, we should support `-config path/to/file`.