[SYSTEMML-558] Fix frame copy/indexing w/ different schemas, tests
This patch generalizes the frame copy primitive to allow different
schemas for source and target frames. There are multiple usecases like
reading of frames in binary block, leftindexing, and frame reuse where
this flexibility leads to much better robustness. With regard to
performance, we still provide a fast special case if schemas of source
and target match (per column).