[SYSTEMML-641] Performance core block matrix mult (colwise par dense)
This patch generalizes the dense-dense matrix-matrix block
multiplication by the ability to parallelize over column blocks of the
rhs matrix. We apply this for small lhs and very wide right hand-side
matrices if the left hand-side fits into l2 cache. This leads to a much
better cache behavior because each thead only scans portions of the
large rhs matrix.