[SYSTEMML-652] Fix function inlining in unsupported expressions, tests
Function calls in expressions (commonly left indexing expressions) are
not supported and should result in an language exception. However, with
function inlining, the target was always assumed to be a data
identifier, which silently changed indexing expressions to assignments
leading to potentially incorrect results. This patch prevents function
inlining for unsupported target expressions for consistent error
handling.