[SYSTEMML-631] Improve handling of namespace conflicts.
This patch with new tests changes the way namespaces are tracked internally (i.e., uses qualified script file path) to prevent collisions. Previously, functions would be lost if the same namespace was specified when importing different scripts which resulted in error being reported as function undefined in namespace. With this change, 'source(file1) as namespace1' in one script does not interfere with 'source(file2) as namespace1' in another script. Note an error is reported at source statement if same namespace specified for different scripts within the same file only.

Closes #133.