[SYSTEMML-676] Improve PyDML Slicing
Currently, slicing correctly uses inclusive lower and exclusive upper bounds. However, a really useful piece of slicing syntax we are currently missing from PyDML is the ability to have implicit lower or upper bounds, i.e., `X[:3,]` should return all rows up to (but not including) 3. This means that the implicit lower bound is 0. Similarly, `X[2:,]` should return all rows starting at row 2 (0-based), which implies that the upper bound is equal to the number of rows of `X`.

This change updates the grammar to allow these cases, and more generally, `X[:,:]`, with both row and column implict slicing as in NumPy.

Closes #148.