[SYSTEMML-753][SYSTEMML-633] Extended loop update-in-place left indexing
This patch generalizes the existing rewrite for marking loop left
indexing as update-in-place. We now additionally tolerate (1) statement
blocks where the respective variable is not used as all, and (2) safe
nrow/ncol operations in the same DAG. Further, the for/while runtime
update-in-place is now also used by parfor, which is useful if the
parfor optimizer cannot pin the variables for update in place due to
unknown sizes, e.g., with complex UDF functions.

Furthermore, this patch also makes smaller improvements to CSR in-place
indexing (ensure exponential allocation in all cases, efficiency set
index range, no special append on copy dense-sparse if not MCSR).

Finally, the explain util has been fixed to output the function call DAG
just once (not on every function) and output update-in-place variables
for parfor loops as well.