[SYSTEMML-630] Fix robustness csv frame readers (count num columns)
In corner cases, our distributed csv write produces empty partitions and
hence empty csv file splits. Depending on the ordering of splits, this
causes problems when determining the number of columns on csv read with
a single row only. This patch introduces a more robust approach for
computing the number of columns on csv read, where we read split headers
until we find a row with ncol>0.