[SYSTEMML-618] SystemML-NN: Simplying the execution of the examples.
Simplying the execution of the examples by introducing a soft link inside the `examples` folder pointing to the actual `nn` folder, and moving the notebooks inside the `examples` folder. Now, all example scripts can be run independently of the enclosing `examples` folder, and users should invoke these scripts from the directory in which they are located.