[SYSTEMML-769] Adding memoryless loop-based convolution operator that is useful in low-memory situation
The initial set of experiments on Lenet have show that there is a
performance degradation due to Cache RLS for im2col approach for large
number of images and small memory budget size (as example: with a 20g JVM,
the overhead is additional 1000 seconds). This degradation is not observed
when we increases the  memory budget (as an example: with 80g JVM).

Additional features:
1. Moved setting of nnz from instruction to LibMatrixDNN
2. Improved the degree of parallelism of ConvTask

TODOs:
- Improve the performance of LibMatrixDNN.doConv2d_Backward_Filter. Since
  this was not in top 10 heavy hitter, I have left this for subsequent PR.