[SYSTEMML-618][SYSTEMML-807][SYSTEMML-808] Adding new LSTM and RNN layers.
This adds two new recurrent neural net layers: LSTM, RNN.  Generically,
recurrent neural nets operate over sequences of examples, and at each
timestep, a single example and the output of the network at the previous
timestep are both passed in as inputs.  The `rnn` layer implements this
generic algorithm, and an LSTM (Long Short Term Memory cell) is simply
an RNN with a slightly more complex algorithm that maintains an
additional internal cell state.  Generally, the LSTM is preferred to the
vanilla RNN.