[SYSTEMML-865] Add Matrix Wrapper to Python MLContext
This exposes the Matrix type that is present in the new Java MLContext to the new Python MLContext by allowing the Java side to return all of the correct types via a simple call to get() on a MLResults object, including returning the existing Matrix type. On the Java side, the new Matrix type can already be used for future SystemML scripts, and contains methods for converting to other types, such as a Spark DataFrame. On the Python side, this simply creates a wrapper around this Java Matrix object. Then, the user can either use this Matrix object in a future SystemML script, or can convert it to a PySpark DataFrame with toDF().

Closes #217.