[SYSTEMML-889] Remove MLContext Java input output methods
This removes the `Script.input` and `Script.output` methods on the JVM
side, and hacks the Python side so that it is able to call the
`Script.in` method on the JVM side.  This means that a Scala/Java user
would use the clean `in(...)` and `out(...)` syntax, while a Python
user would use the `input(...)` and `output(...)` syntax.