[SYSTEMML-869] Fix implicit cleanup of mlcontext input variables, tests
Depending on the script control structure (e.g., multiple statement
blocks), passed intermediates from one script to another script might be
cleaned up as they appear to be no longer used. However, this leads to
destroyed intermediates that a user might still want to pass to other
scripts. Hence, we now disable the cleanup of any mlcontext input
variables as this should be in the hand of the user. 

Furthermore, this patch also includes some minor cleanups of the
monitoring api.