[SYSTEMML-569] Scalability transformencode (memory-efficient build)
The scalability of transformencode with recoding depends on the number
of distinct items in the input frame. The larger the number of distinct
items, the larger the temporary partial recode maps and the larger the
output per partition (all of which need to fit into the working memory).
For a scenario of 1G x 5 w/ double schema and 1Mx5 distinct items, we
previously ran into out-of-memory errors.

This patch makes various modifications: 

(1) New custom partial build, which avoids code assignments (long
objects) and unnecessary string conversions of schema-specific values.
This is beneficial for performance (string conversion and reduce garbage
collection) and more importantly reduce the memory footprint (no long
objects and smaller tokens on non-string schemas).

(2) Global distinct/group/write on schema-specific objects instead of
string tokens. This again reduces the memory footprint on the map task
output and group output but also reduces shuffle.  

With these changes, we're now able to execute the aforementioned
scenario in 782s. However, it also generally improved performance: for
example, on a scenario of of 1G x 5 w/ double schema and 100Kx5 distinct
items, improved runtime from 298s to 158s.