[SYSTEMML-959] Performance spark grouped aggregate (single block agg)
Similar to spark unary aggregates and various other instructions, this
patch introduces runtime-level alternatives for single-block and
multi-block aggregation. For example, naive bayes over a 10Mx1k, dense
input (80GB) improved from 64 to 53s. 

Furthermore, this also includes a correctness fix of deep copying matrix
blocks on the creation of combiner blocks which is a pre-condition for
subsequent update in-place if the blocks are consumed by multiple
operations.