[SYSTEMML-965] Performance reduce-all rdd operations (in-place fold)
The standard binary 'reduce' rdd operation requires to deep copy the
left input block to the result as it is unknown if the inputs are
partial aggregates or original blocks, which are potentially consumed by
other operations. We now use 'fold' with explicit zero element to avoid
unnecessary deep copies (object allocation and copy). On dense and
sparse linregds scenarios this showed improvements up to 6%. Note that
this change applies to all instructions relying on the core primitives
sumStable or aggStable as well as centralMoment and covariance.