[SYSTEMML-958] Improved multilogreg algorithm script (known size of Y)
The size of the modified label vector Y (as indicator matrix with one
column per class label) is data dependent. Despite worst-case estimates
and dynamic recompilation, this leads to unnecessary distributed
operations for the table operation that creates the indicator matrix.
This script change makes a slight modification to this preprocessing,
which allows dynamic recompilation to exactly infer the output
dimensions. On a scenario of 100M x 10, 2 classes, and 15GB driver this
led to an improvement from 360s to 230s end-to-end.