[SYSTEMML-968] Fix buffer pool integration of frames (evict, serialize)
There was a basic integration of frames into the buffer pool. However,
the in-memory and serialized size was always reported as 1, which
effectively avoided evictions for frames. Accordingly, with sufficiently
large frame intermediates, out-of-memory errors occurred.

This patch now fully integrates frames into the buffer pool, which
covers the following changes:

(1) Implementation of frame block getInMemorySize and
getExcactSerializedSize as well as the decision on shallow serialize.

(2) Implementation of writeUTF in our CacheDataOutput and
FastBufferedOutputStream, which avoids redundant copies compared the
default approach of new DataOutputStream(new BufferedOutputStream(out)).

(3) Handling of deserialization and read of frame blocks on restore,
which so far only handled matrix blocks.  

(4) Variety of testcases with serialization/deserialization, write/read,
dense/sparse, and a number of different schemas.