[SYSTEMML-1004] New spark tsmm2 matrix mult operator (compiler/runtime)
If the number of columns exceeds the blocksize for expressions like t(X)
%*% X, we can no longer compile a spark/mr tsmm but use cpmm as a
fallback. Since cpmm shuffles both inputs, this had huge performance
impact. This patch introduces a new tsmm2 operator, which applies if the
number of columns is less or equal than 2x the blocksize, by using two
passes over the input (which is in comparison to shuffle significantly
faster). On our perftest scenario of 10Mx1K, icpt={1|2} for LinregDS,
this improved performance from ~300 to 95s (including spark context
creation and initial read; so the actual improvement for this operation
is even larger).