[SYSTEMML-1011] Performance sparse append cbind (allocate rows once)
So far, our sparse append cbind operations used an approach of append
both lhs and rhs to the output. Implicitly, this created a deep copy of
the lhs and reallocated rows (with log N steps) for rhs. We now
explicitly determine the nnz per row and allocate the output sparse row
once. Even on appending a column vector this consistently improved
performance by 10%. However, the performance improvements increase as
the size of the rhs increases.