[SYSTEMML-951] Performance spark right indexing (no agg, block lookup)
This patch adds two improvements to the spark right indexing instruction
for the use case of batch extraction out of large input matrices. 

(1) Scan-based indexing w/o aggregation: If the indexing range is block
aligned, i.e., each dimension of the indexing range starts at block
boundaries or touches only a subset of a single block, there is no need
to merge partial blocks. We now determine this characteristics at
runtime and avoid the unnecessary mergeByKey.

(2) Single-block lookups: If the indexing range touches only a single
block, we now perform this indexing via a single rdd lookup and
subsequent in-memory slicing. For in-core matrices or without data
partitioner this does not hurt performance, whereas for out-of-core
matrices, this gives a huge improvement as we only need to scan a single
partition instead of the entire dataset.

For further details, on multi-block lookups, see
https://issues.apache.org/jira/browse/SYSTEMML-951.