[SYSTEMML-951] Efficient spark multi-block indexing (partition pruning)
This patch introduces efficient multi-block spark right indexing. We
leverage Spark's PartitionPruningRDD to prune partitions by determining
the set of required partitions based on necessary block indexes and
wrapping the PartitionPruningRDD (which acts as a filter) around the
input. In contrast to data-query joins, this changes the asymptotic
behavior on out-of-core datasets as only required partitions are read
from secondary storage. 

This code path is applied if (1) the input has a hash partitioner, (2)
the input is an out-of-core dataset, (3) the indexing range is block
aligned, and (4) the output fits twice in local memory. Note that
restrictions (2)-(4) could be removed but this conservative approach
ensures that it will never be counter-productive or run into local
memory issues in the preparation phase. For the use case and details,
see https://issues.apache.org/jira/browse/SYSTEMML-951.