[SYSTEMML-1029] Fix performance CSR update-in-place left indexing, test
Sparse update-in-place matrices are represented in CSR (instead of MCSR)
in order to enable shallow serialize because the CSR in-memory size is
almost equivalent its serialized representation. This is important
because otherwise, the O(n) serialization costs would eliminate the
benefits of update-in-place. 

However, column-wise CSR updates led to severe performance issues as
each row update triggered shifting of all subsequent values leading to
O(n^2) shifting costs. This patch adds low-level CSR update routines
which require at most O(n) shifting costs for this worst-case access
pattern (which is still faster than serialization or copy). Note that
the full benefits of append-only update-in-place remain unchanged.