[SYSTEMML-951] Fix Block Alignment Check For Multi-Block Index Pruning
In `isMultiBlockLookup` we check that the index range is block aligned
with `OptimizerUtils.isIndexingRangeBlockAligned`.  In that method, we
aim to grab the rows & cols per block.  However, instead we end up
grabbing the number of row blocks & column blocks.  This patch changes
this logic to correctly compute the rows & columns per block.