[SYSTEMML-1045] Fix CSR block growth rate on resize (update-in-place)
So far, we used an aggressive 2x growth rate for CSR sparse
representations. In comparison to MCSR, this significantly affects
temporary memory requirements as this applies to the entire matrix not
just a single row. For update-in-place left indexing, the growth is
covered by the input + output memory requirements, which are roughly 2x
the input but not 3x as required for a rate of 2x. We now use a
conservative 1.1x rate. Furthermore, this patch also cleans up some
unnecessary redundancy of the related resizing logic and fixes incorrect
size computation (for small initial capacities).