[SYSTEMML-1128] DML utility script to shuffle input dataset
Since this is a frequently used operation, I thought it would be nice to
have this script in our utils. In fact, sklearn too has this
functionality in its util: sklearn.utils.shuffle

As an FYI, this script took 4.026 sec to shuffle 658 MB binary input
data of shape [3444983, 10].

Also, removed cbind.dml as we now provide this functionality through
built-in function.

Closes #293.