[SYSTEMML-1148] Explicitly set value types for MLContext scalar inputs
Not explicitly setting the scalar value type of a basic input type
through the MLContext API can cause issues since the value type can
default to double. For example, "print(!b)" where b is a boolean can
throw an error since the NOT operation can be performed on a boolean
but not a double.

Closes #314.