[SYSTEMML-1140] Performance binary sparse-dense mult/div (pre-allocate)
This patch improves the performance of sparse-dense element-wise
multiply and divide by pre-allocating the sparse output rows once with
the size of the sparse input rows, which likely comes close to the
output size given the dense right-hand-side. 

Note that this patch does not fix all issues of [SYSTEMML-1140] but is a
related backlog item that we've encountered during the SPOOF
experiments. 

The performance improvements are moderate at around 25% but tend to
increase with skew of sparse row sizes. On mnist60k, div sparse-dense,
the improvement was from 391ms to 296ms.