[SYSTEMML-1251] Fix recompile-once recursive functions, cleanup, tests
This patch fixes our inter-procedural analysis, disallowing to mark
directly or indirectly recursive functions for recompile-once because
recompile-once can lead to OOMs and even incorrect results in the
context of recursion. Recompile-once functions are recompiled on entry
with the given input statistics in order to avoid repeated recompilation
in loops, which is often unnecessary knowing the size of function
inputs. However, with recursive functions it is not that easy. A subcall
recompiles the plan again, also modifying the remaining plan for the
calling function - if both recursion levels work on different sizes or
compile literals into the plan, the consequences are disastrous. 

Since multiple places such as EXPLAIN and IPA require information about
the function call graph, this patch introduces a clean abstraction, the
FunctionCallGraph, that captures this information once and simplifies
related compiler passes.

Furthermore, this patch also includes a fix for our -stats tool to
properly reset the number and time of function recompilations.