[SYSTEMML-1273] Performance spark right indexing (aligned, w/o agg)
This patch adds a special case for spark right indexing without
aggregation, where a single block is either directly passed through to
the output or sliced to a single output block. The main application is
aligned slicing of entire row batches. The change also includes specific
tests for this special case.